pyproject.toml
[tool.poetry]
name = "solana_sandbox"
version = "0.1.0"
description = ""
authors = ["Lukasz Laszko <lukasz@fairtide.co>"]

[tool.poetry.dependencies]
click = "^8.0.4"
python = "^3.9"
python-box = "^5.4.1"
loguru = "^0.6.0"
solana = "^0.21.0"

[tool.poetry.dev-dependencies]
black = ""
pytest = "^7.0.1"
pyfiglet = ""

[tool.black]
--skip-string-normalization = "yes"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
